+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/CRUD/Read/Excel/ExcelExportTest.cs
@@ -93,7 +93,7 @@
         {
             ActODataService(args =>
             {
-                DataServiceProvider.DataService = args.DataService;
+                DataService = args.DataService as SQLDataService;
                 args.Token.Functions.Register(new Func<QueryParameters, string, Страна[]>(FunctionExportExcel));
 
                 // Create objects and put them in the database.
@@ -104,7 +104,7 @@
                     countries[i] = new Страна { Название = string.Format("Страна №{0}", i) };
                 }
 
-                args.DataService.UpdateObjects(ref countries);
+                DataService.UpdateObjects(ref countries);
                 string requestUrl = string.Format(
                     "http://localhost/odata/{0}?{1}",
                     "FunctionExportExcel(entitySet='Странаs')",
@@ -118,6 +118,8 @@
             });
         }
 
+        private SQLDataService DataService { get; set; }
+
         /// <summary>
         /// Функция подготавливающая данные для экспорта в Excel. Для правильной работы необходимо, чтобы в декларации был указан реальный тип возвращаемых значений.
         /// Не подходит указание типа DataObject.
@@ -125,12 +127,11 @@
         /// <param name="queryParameters"></param>
         /// <param name="entitySet"></param>
         /// <returns></returns>
-        private static Страна[] FunctionExportExcel(QueryParameters queryParameters, string entitySet)
+        private Страна[] FunctionExportExcel(QueryParameters queryParameters, string entitySet)
         {
-            SQLDataService dataService = DataServiceProvider.DataService as SQLDataService;
             Type type = queryParameters.GetDataObjectType(entitySet);
             LoadingCustomizationStruct lcs = queryParameters.CreateLcs(type);
-            Страна[] dobjs = dataService.LoadObjects(lcs).Cast<Страна>().ToArray();
+            Страна[] dobjs = DataService.LoadObjects(lcs).Cast<Страна>().ToArray();
             return dobjs;
         }